+++ aws-java-sdk-kinesisanalyticsv2/src/main/java/com/amazonaws/services/kinesisanalyticsv2/AmazonKinesisAnalyticsV2Async.java
@@ -377,8 +377,7 @@ public interface AmazonKinesisAnalyticsV2Async extends AmazonKinesisAnalyticsV2 
 
     /**
      * <p>
-     * Creates and returns a URL that you can use to connect to an application's extension. Currently, the only
-     * available extension is the Apache Flink dashboard.
+     * Creates and returns a URL that you can use to connect to an application's extension.
      * </p>
      * <p>
      * The IAM role or user used to call this API defines the permissions to access the extension. After the presigned
@@ -409,8 +408,7 @@ public interface AmazonKinesisAnalyticsV2Async extends AmazonKinesisAnalyticsV2 
 
     /**
      * <p>
-     * Creates and returns a URL that you can use to connect to an application's extension. Currently, the only
-     * available extension is the Apache Flink dashboard.
+     * Creates and returns a URL that you can use to connect to an application's extension.
      * </p>
      * <p>
      * The IAM role or user used to call this API defines the permissions to access the extension. After the presigned